Move run function to the library from the app
